% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVM.TEST.R
\name{CVM.TEST}
\alias{CVM.TEST}
\title{A two-sample Cramer-von Mises test for equal distributions as a permutation test using R.}
\usage{
CVM.TEST(x, y, R)
}
\arguments{
\item{x}{the data from the first distribution}

\item{y}{the data from the second distribution}

\item{R}{the number of permutation}
}
\value{
a statistic for the test \code{statistic}

a p-value for the test \code{p_value}
}
\description{
A two-sample Cramer-von Mises test for equal distributions as a permutation test using R.
}
\examples{
\dontrun{
attach(InsectSprays)
x1 <- as.vector(count[spray == "A"])
x2 <- as.vector(count[spray == "B"])
x3 <- as.vector(count[spray == "C"])
x4 <- as.vector(count[spray == "D"])
x5 <- as.vector(count[spray == "E"])
x6 <- as.vector(count[spray == "F"])
CVM.TEST(x1, x2, 299)
CVM.TEST(x3, x4, 299)
}
}
