% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{f18}
\alias{f18}
\title{A simple statistical function generator using Rcpp.}
\usage{
f18()
}
\value{
a density fucntion \code{y1}.

a distribution fucntion \code{y2}.

a quantile fucntion \code{y3}.
}
\description{
A simple statistical function generator using Rcpp.
}
\examples{
\dontrun{
f18()
}
}
