% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q.test_interval.R
\name{q.test_interval}
\alias{q.test_interval}
\title{A quantile.test with confidence interval(two sided) using R}
\usage{
q.test_interval(
  x,
  xstar = 0,
  quant = 0.5,
  alternative = "two.sided",
  conf.level = NULL
)
}
\arguments{
\item{x}{the sample data}

\item{xstar}{the pre-specified points}

\item{quant}{the pre-specified quantile}

\item{alternative}{the test types:two.sided&left.sided&right.sided}

\item{conf.level}{the confidence level for two.sided test}
}
\value{
The first statistic of quantile.test \code{T1}

The second statistic of quantile.test \code{T2}

The p-value of quantile.test \code{p.value}

the confidence level of quantile.test with two.sided \code{conf.level}

the confidence interval of quantile.test with two.sided \code{conf.interval}
}
\description{
A quantile.test with confidence interval(two sided) using R
}
\examples{
\dontrun{
x<-c(199,213,202,193,174,166,248,189,233,195,160,212,176,231,185)
q.test_interval(x,xstar=193,quant=0.75,conf.level=0.9,alternative="two.side")
q.test_interval(x,xstar=193,quant=0.75,conf.level=0.9,alternative="left.side")
q.test_interval(x,xstar=193,quant=0.75,conf.level=0.9,alternative="right.side")
}
}
