\name{StatComp22072-package}
\alias{StatComp22072-package}
\alias{StatComp22072}
\docType{package}
\title{
  This package provides homework for 'Statistics Computing' course and several R and Rcpp function for display.
}
\description{
  Homework for 'Statistics Computing' course, R version of the Cramer-von Mises test and the quantile.test, Rcpp version of gibbs sampler are provided in this package(\code{CVM.TEST} and \code{q.test_interval} and \code{f18})
}
\author{
Guohui Jiang

Maintainer: Guohui Jiang <13625673575@126.com>
}
\keyword{ package }
